use LF instead of CRLF
What is changed:
- replaced CRLF with LF by running `dos2unix *`
- configure editorconfig to use LF

It is always recommended to use LF rather than CRLF in programming
world, we should keep this practice in Obsidian plugins.

For anyone who wants to use CRLF locally, you can set
`git config core.autocrlf` to true. For further information,
this article is a good start:
https://www.aleksandrhovhannisyan.com/blog/crlf-vs-lf-normalizing-line-endings-in-git/